Remove unneded server js, moving component to different file
